refactor: MemberInfo 컴포넌트명을 MemberAddressInfo로 변경
